+++ device/x86/TopK/top_k.c
@@ -34,7 +34,7 @@ int eval(BUFFER_INFO_S *params, BUFFER_INFO_S *inputs, BUFFER_INFO_S *outputs) {
         rem_buf_ptr += (ifmap_need_buf_size + 31) & (~32);
         rem_buf_size -= (ifmap_need_buf_size + 31) & (~32);
     } else {
-        LOG_ERR("remaining buf size is %d, but need buf size is %d", rem_buf_size, ifmap_need_buf_size);
+        LOG_ERR("remaining buf size is %d, but need buf size is %d", (int32_t)rem_buf_size, (int32_t)ifmap_need_buf_size);
     }
 
     memcpy(ifmap_ptr, input_ptr, in_elem_size * sizeof(float ));
